+++ modules/misc.js
@@ -1,9 +1,11 @@
 import * as MAESTRO from "./config.js";
 import * as Playback from "./playback.js";
+import { addFolderImportButton } from "./playlist-import.js";
 
 
 export function _onRenderPlaylistDirectory(app, html, data) {
     _addPlaylistLoopToggle(html);
+    addFolderImportButton(html);
 }
 
 export class MaestroConfigForm extends FormApplication {